Start of a GetFeedback test using FaultFinder
Added the "ElementFilter" class to create the iterator as it is required.

Added a temporary "toJson" function to the "CorniTransformation" class since it'll be necessary for the probe.

Added a working "toJson" function to the "'MoveBottomTransformation" class.

Started to create a "GetFeedbackTest" class to test draft of a "process" function using FaultFinder.